+++ workspaces/gamut-storybook/stories/Badge.stories.js
@@ -1,9 +1,0 @@
-import { Badge } from 'gamut';
-import React from 'react';
-
-export default {
-  component: Badge,
-  title: 'Component/Badge',
-};
-
-export const badge = () => <Badge>New</Badge>;